week7 미션

+++ src/main/java/umc_8th/spring/Application.java
@@ -23,7 +23,7 @@ public class Application {
 
 			// 파라미터 값 설정
 			String name = "요아정";
-			Float score = 4.0f;
+			Double score = 4.0;
 
 			// 쿼리 메서드 호출 및 쿼리 문자열과 파라미터 출력
 			System.out.println("Executing findStoresByNameAndScore with parameters:");
